tools/vector_build/requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
numpy>=1.26.0
tqdm>=4.66.0
